+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@
 
 - Adjusted mutability of getters in Config.
 - Added hide/show button operations to the sample app.
+- Removed event handling for widgets that are not visible (#103)
+- Enable/Disable state on Widget implemented (#39)
 
 ## 0.2.6